# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "KnxMonitor.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Telegrams": {
      "get": {
        "tags": [
          "Telegrams"
        ],
        "summary": "Keyset-paginated telegram history (newest first).",
        "description": "Reads the database hot tier only — the day archive on disk is served by the archive/* endpoints. Filters: from/to, address (destination group address), source (individual address), type or types (comma-separated, OR-combined and taking precedence over type) and q (free-text LIKE across source, destination, decoded value, GA name and DPT). Paging is keyset-based over (Timestamp, Id), not offset-based: pass the nextCursor of the previous response as cursor to get the following page. The cursor is opaque; a malformed one yields 400. pageSize defaults to 100 and is clamped to 1..1000. order=asc sorts oldest first, any other value newest first — the sort direction must stay the same while paging, because the cursor is interpreted relative to it. from/to are compared against the stored UTC timestamps without any conversion (unlike the chart endpoints), so pass UTC values.",
        "parameters": [
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "To",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Address",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Source",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Types",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of telegrams plus hasMore and nextCursor; nextCursor is null on the last page."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Telegrams"
        ],
        "summary": "Permanently deletes the entire telegram history (the DB hot-tier). Irreversible.",
        "description": "Deletes every row in one statement — the request carries no filter, so there is no way to clear only part of the history. The SQLite WAL is checkpointed afterwards so the file shrinks. Archive files on disk are not touched and stay downloadable.",
        "responses": {
          "200": {
            "description": "An object with a single deleted property: the number of rows removed."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Telegrams/count": {
      "get": {
        "tags": [
          "Telegrams"
        ],
        "summary": "Number of telegrams matching a filter.",
        "description": "Takes the same query parameters as GET /api/telegrams, but only the filtering ones have an effect — cursor, pageSize and order are ignored. Counts the database hot tier only.",
        "parameters": [
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "To",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Address",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Source",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Types",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An object with a single count property."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Telegrams/export": {
      "get": {
        "tags": [
          "Telegrams"
        ],
        "summary": "Server-streamed CSV export over the full filtered DB range.",
        "description": "Accepts the same filters as GET /api/telegrams, but exports every matching row: cursor and pageSize are ignored and there is no row limit. Rows are always written chronologically (oldest first), regardless of order. The response is streamed while the query is still running, so the body starts arriving before the row count is known and no Content-Length is sent. The header line is always written, even when nothing matches. Encoding is UTF-8 without BOM; fields that contain a comma, quote or line break are quoted with doubled inner quotes. The filename is knx-telegrams-{UTC timestamp}.csv. Columns: Timestamp, SourceAddress, DestinationAddress, GroupAddressName, DatapointType, MessageType, Value, ValueDecoded. Value is the raw payload as uppercase hex without separators. Timestamp is the UTC value in ISO-8601 round-trip format but carries no zone designator — unlike the JSON responses, which end in Z. GroupAddressName and DatapointType are empty when the telegram has no linked group address.",
        "parameters": [
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "To",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Address",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Source",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Types",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Only csv is supported; any other value yields 400.",
            "schema": {
              "type": "string",
              "default": "csv"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A streamed text/csv attachment."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Telegrams/archive/days": {
      "get": {
        "tags": [
          "Telegrams"
        ],
        "summary": "Lists the days available in the cold-tier archive, newest first.",
        "description": "The archive holds one NDJSON file per UTC day, written independently of the database hot tier and unaffected by its retention or by DELETE /api/telegrams. Finished days are gzipped (compressed: true), the current day stays plain and is still being appended to, so its size grows between calls. sizeBytes is the size on disk, i.e. the compressed size for gzipped days. Archiving is opt-in via the recording settings, so the list is empty as long as nothing has been archived.",
        "responses": {
          "200": {
            "description": "One entry per day with date (YYYY-MM-DD), fileName, sizeBytes and compressed."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Telegrams/archive/{date}": {
      "get": {
        "tags": [
          "Telegrams"
        ],
        "summary": "Downloads the raw archive file of one UTC day.",
        "description": "Serves the gzipped file (application/gzip) when it exists and falls back to the plain, still-open file of the current day (application/x-ndjson); 404 when neither is present for that day. The payload is NDJSON — one JSON object per line with the keys t (UTC timestamp), s (source address), d (destination address), mt (message type), r (raw value) and v (decoded value, omitted when there is none). The current day is served while the writer still holds it open, so it can be missing the last few seconds — the writer flushes on an interval.",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "Archive day as YYYY-MM-DD (UTC). Any other shape yields 400.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Access-Token aus POST /api/Auth/login (ohne \"Bearer \"-Präfix einfügen).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}